nats_worker/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
fast-depends==2.4.12
faststream==0.5.35
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
nats-py==2.9.0
pydantic==2.10.6
pydantic_core==2.27.2
python-telegram-bot==21.11.1
sniffio==1.3.1
typing_extensions==4.12.2